### 2017-07-05 #130 @cboettig
codemetar package
Codemeta defines a 'JSON-LD' format for describing software metadata.
    This package provides utilities to generate, parse, and modify codemeta.jsonld
    files automatically for R packages.

```
Package: codemetar
Title: Generate CodeMeta Metadata for R Packages
Version: 0.1.0
Authors@R: person("Carl", "Boettiger", role=c("aut", "cre", "cph"), email="cboettig@gmail.com", comment="http://orcid.org/0000-0002-1642-628X")
Description: Codemeta defines a 'JSON-LD' format for describing software metadata.
    This package provides utilities to generate, parse, and modify codemeta.jsonld
    files automatically for R packages.
Imports: jsonlite (>= 1.3),
    jsonld,
    git2r,
    devtools,
    stringi
Suggests: testthat,
    jsonvalidate,
    covr,
    httr,
    xml2
keywords: metadata, codemeta, ropensci, citation, credit
affiliation: https://ropensci.org
relatedLink: https://codemeta.github.io/codemetar

```

-   URL for the package (the development repository, not a stylized html page): https://github.com/codemeta/codemetar

Package for creating and working with scientific software metadata

Academic researchers looking to create metadata for their software

Nope

A few lines of code handle exceptional cases that are difficult to cover in unit tests.  Otherwise no there should be no outstanding `goodpractice` flags in the repo.

Perhaps anyone interested in metadata/archiving or just in messing around with json, maybe:

@amoeba @gothub @jennybc @sckott @stephlocke @jeroen @mfenner @o2r-project
