### 2017-07-06 e34e16d
update path based test for appveyor
