+++ ropensci-onboarding.md
@@ -79,7 +79,7 @@ Confirm each of the following by checking the box.  This package:
 
 - [x] Do you intend for this package to go on CRAN?
 - [x] Do you wish to automatically submit to the [Journal of Open Source Software](http://joss.theoj.org/)? If so:
-    - [ ] The package contains a [`paper.md`](http://joss.theoj.org/about#paper_structure) with a high-level description in the package root or in `inst/`.
+    - [x] The package contains a [`paper.md`](http://joss.theoj.org/about#paper_structure) with a high-level description in the package root or in `inst/`.
     - [ ] The package is deposited in a long-term repository with the DOI:
     - (*Do not submit your package separately to JOSS*)
 
