### 2017-07-07 #130 @maelle
Thanks for your submission @cboettig ! 😸

Running `devtools::check()` I get

```r
Quitting from lines 118-126 (codemeta-parsing.Rmd)
Error: processing vignette 'codemeta-parsing.Rmd' failed with diagnostics:
could not find function "as_tibble"
Execution halted
Error: Command failed (1)
```

Which I cannot make sense of yet.

Note: I'll replace this comment with the usual editor checks once this gets sorted out. 😄
