### 2017-07-07 #130 @maelle
For info currently running the checks again after replacing all `as_tibble` with `as_data_frame` (`dplyr` wrapper for `as_tibble`).
