### 2017-07-07 #130 @maelle
Regarding the test skipped on Appveyor I see why you're skipping it, the code doesn't work on Windows. 😉 If it's an important test, why not write a version using if's? 🤔
