### 2017-07-07 #130 @maelle
Thanks for your submission @cboettig ! 😸  Looking for reviewers now. I also have two comments about one vignette and one test below.

### Editor checks:

- [ x] **Fit**: The package meets criteria for [fit](policies.md#fit) and [overlap](policies.md#fit)
- [ x] **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service.
- [ x] **License:** The package has a CRAN or OSI accepted license
- [ x] **Repository:** The repository link resolves correctly

---

#### Editor comments

* Nothing interesting from `goodpractice::gp()` 😉

* Running `devtools::check()` I get

```r
Quitting from lines 118-126 (codemeta-parsing.Rmd)
Error: processing vignette 'codemeta-parsing.Rmd' failed with diagnostics:
could not find function "as_tibble"
Execution halted
Error: Command failed (1)
```

unless I replace all `as_tibble` with `as_data_frame` (`dplyr` wrapper for `as_tibble`).

* Regarding the test skipped on Appveyor (test-cloned-repo.R) I see why you're skipping it, the code doesn't work on Windows. 😉 If it's an important test, why not write a version using if's? 🤔
---

Reviewers:
Due date:
