### 2017-07-07 e961961
update appveyor badge
