### 2017-07-07 #130 @cboettig
@maelle what version of dplyr are you using? `as_tibble()` is (once again) exported from `dplyr` as of the `0.7.0` release: https://github.com/tidyverse/dplyr/blame/master/NEWS.md#L48-L49

Re the cloned-repo test on Windows, the [error I see on Appveyor](https://ci.appveyor.com/project/cboettig/codemetar/build/1.0.1) comes from the `devtools::use_build_ignore()` function, and I couldn't find examples from any [unit tests in the devtools package for that function](https://github.com/hadley/devtools/search?utf8=%E2%9C%93&q=use_build_ignore&type=) so I'm a bit at a loss.  Must have something to do with how Windows handles file paths, but not sure what.  Does that `devtools` function work on windows with the default argument to `pkg = "."` for the path?
