### 2017-07-07 #130 @maelle
Ill look on Sunday or Monday but maybe add 0.7.0 as minimal version in DESCRIPTION?

For me the first error was bc of the file path, the beginning worked when omitting the / at the end but I'll look again in a few days and hopefully make a PR or suggestions. Windows user at your service! 😉
