### 2017-07-07 e7e87b1
indicate dplyr version
