### 2017-07-07 1a29a14
update codemeta record with dplyr version ;-)
