### 2017-07-09 8e09357
Tries to handle Windows
