+++ tests/testthat/test-cloned-repo.R
@@ -10,18 +10,20 @@ testthat::test_that("we can generate codemeta
 
     git2r::clone("https://github.com/codemeta/codemetar",
                  "codemetar_copy", progress = FALSE)
-    write_codemeta("codemetar_copy/", "test.json")
-    expect_true(codemeta_validate("test.json"))
 
-    f <- "codemetar_copy"
+    if(as.character(Sys.info()['sysname']) == "Windows"){
+      write_codemeta("codemetar_copy", "test.json")
+      expect_true(codemeta_validate("test.json"))
 
-    test_that("git utils", {
-      x <- uses_git(f)
-      testthat::expect_true(x)
-      x <- guess_github(f)
-      x <- github_path(f, "README.md")
+      f <- "codemetar_copy"
 
-    })
+      test_that("git utils", {
+        x <- uses_git(f)
+        testthat::expect_true(x)
+        x <- guess_github(f)
+        x <- github_path(f, "README.md")
+
+      })
 
       guess_ci(file.path(f, "README.md"))
       guess_devStatus(file.path(f, "README.md"))
@@ -31,10 +33,38 @@ testthat::test_that("we can generate codemeta
       guess_fileSize(f)
       unlink("codemetar-*.tar.gz")
 
-    unlink("test.json")
-    system("rm -rf codemetar_copy")
-    #file.remove(list.files("codemetar_copy", recursive = TRUE))
-    #unlink("codemetar_copy")
+      unlink("test.json")
+      file.remove(list.files("codemetar_copy", recursive = TRUE))
+      unlink("codemetar_copy")
+    }else{
+      write_codemeta("codemetar_copy/", "test.json")
+      expect_true(codemeta_validate("test.json"))
+
+      f <- "codemetar_copy"
+
+      test_that("git utils", {
+        x <- uses_git(f)
+        testthat::expect_true(x)
+        x <- guess_github(f)
+        x <- github_path(f, "README.md")
+
+      })
+
+      guess_ci(file.path(f, "README.md"))
+      guess_devStatus(file.path(f, "README.md"))
+
+      guess_readme(f)
+      guess_releaseNotes(f)
+      guess_fileSize(f)
+      unlink("codemetar-*.tar.gz")
+
+      unlink("test.json")
+      system("rm -rf codemetar_copy")
+      #file.remove(list.files("codemetar_copy", recursive = TRUE))
+      #unlink("codemetar_copy")
+    }
+
+
   })
 
 
