### 2017-07-09 3c1647a
more Windows stuff
