### 2017-07-09 #130 @maelle
* I think I updated `dplyr` by mistake since writing my comment but I see you updated DESCRIPTION.

* Regarding the test failing on Windows see https://github.com/codemeta/codemetar/pull/19
