### 2017-07-10 4bd8848
Windows fixes? (#19)
* Tries to handle Windows

* more Windows stuff
