### 2017-07-11 6bc8c22
github notes continue
