+++ man/drop_context.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crosswalk.R
+\name{drop_context}
+\alias{drop_context}
+\title{drop_context}
+\usage{
+drop_context(x, json_output = FALSE)
+}
+\arguments{
+\item{x}{a JSON list (from read_json / fromJSON) or json object (from toJSON)}
+
+\item{json_output}{logical, should output be a json object or a list?}
+}
+\value{
+a list or json object with the "@context" element removed
+}
+\description{
+drop context element from json list or json string
+}
