+++ man/add_context.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crosswalk.R
+\name{add_context}
+\alias{add_context}
+\title{add_context}
+\usage{
+add_context(x, context, json_output = FALSE)
+}
+\arguments{
+\item{x}{a JSON list (from read_json / fromJSON) or json object (from toJSON)}
+
+\item{context}{context to be added, in same format as x}
+
+\item{json_output}{logical, should output be a json object or a list?}
+}
+\value{
+a list or json object with "@context" element added
+}
+\description{
+add context element to json list or json string
+}
