+++ NAMESPACE
@@ -3,6 +3,7 @@
 export(codemeta_validate)
 export(create_codemeta)
 export(crosswalk)
+export(drop_context)
 export(write_codemeta)
 importFrom(devtools,build)
 importFrom(devtools,use_build_ignore)
@@ -10,11 +11,13 @@ importFrom(git2r,branches)
 importFrom(git2r,repository)
 importFrom(jsonld,jsonld_compact)
 importFrom(jsonld,jsonld_expand)
+importFrom(jsonlite,fromJSON)
 importFrom(jsonlite,read_json)
 importFrom(jsonlite,toJSON)
 importFrom(jsonlite,write_json)
 importFrom(methods,is)
-importFrom(stats,na.omit)
+importFrom(readr,cols)
+importFrom(readr,read_csv)
 importFrom(stringi,stri_trans_general)
 importFrom(utils,as.person)
 importFrom(utils,citation)
