### 2017-07-12 a18dfa5
update docs, avoid gh suggests
