+++ vignettes/translating.Rmd
@@ -68,98 +68,73 @@ The CodeMeta Crosswalk table seeks to accomplish a very similar goal.  The cross
 
 ### GitHub
 
+Here we crosswalk the JSON data returned as the repository information from the GitHub API: 
+
+```{r eval = FALSE}
+repo_info <- gh::gh("/repos/:owner/:repo", owner = "ropensci", repo = "EML")
+```
+
+```{r include = FALSE}
+## Actually, we'll use a chached copy and not eval the above chunk to avoid a `gh` dependency:
+repo_info <- read_json(system.file("examples/github_format.json", package = "codemetar"))
+```
+
+Let's just take a look at what the returned json data looks like:
+
 ```{r}
-r <- gh::gh("/repos/:owner/:repo", owner = "ropensci", repo = "EML")
-github_meta <- crosswalk(r, "GitHub")
+repo_info
+```
+
+We can crosswalk this information into codemeta just by supplying the column name to the `crosswalk` function.  This performs the same expansion of the metadata in the GitHub context, followed by compaction into the codemeta context.  Note that terms not recognized/included in the codemeta context will be dropped:
+
+```{r}
+github_meta <- crosswalk(repo_info, "GitHub")
 github_meta
 ```
 
+We can verify that the result is a valid codemeta document:
+
+```{r}
+codemeta_validate(github_meta)
+```
 
 
 ## Transforming into other column schema
 
 
+The above transform showed the process of going from plain JSON data into the codemeta standard serialization.  Similarly, we can crosswalk into any of the other columns in the crosswalk table.  To do so, the `crosswalk` function will first expand any of the recognized properties into the codemeta JSON-LD context, just as above.  Unrecognized properties are dropped, since there is no consensus context into which we can expand them.  Second, the expanded terms are then compacted down into the new context (Zenodo in this case.)  This time,
+any terms that are not part of the codemeta context are kept, but not compacted, since they still have meaningful contexts (that is, full URIs) that can be associated with them:
 
 ```{r}
-crosswalk(r, "Zenodo", "GitHub") %>%
+crosswalk(repo_info, "GitHub", "Zenodo") %>%
 drop_context()
 ```
 
+Thus terms that still have a uncompacted prefix like `schema:` or `codemeta:` reflect properties that we could successfully extract from the input data, but do not have corresponding properties in the Zenodo context.  This is the standard behavior of the 
+compaction algorithm: unrecognized fields are not dropped, but it is also not compacted,
+making it accessible only if referenced explicitly.  
 
 
-## DataCite
+## NodeJS example
 
+NodeJS uses a `package.json` format that is very similar to a simple `codemeta.json` file, though it is not Linked Data as it does not declare a context.  Here we crosswalk an example `package.json` file into proper `codemeta` standard.  
 
-We can illustrate using the expansion and compaction process to crosswalk DataCite's metadata schema into other formats, such as the native codemeta format or the Zenodo format.  First, let's start with some DataCite metadata:
-
-
-```{r include=FALSE, eval=FALSE}
-## nope, sadly this is not quite clean json translation, due to repeated element structure
-datacite_list <-
-httr::GET("https://doi.org/10.5281/zenodo.573741", 
-    httr::add_headers(Accept="application/vnd.datacite.datacite+xml")) %>% 
-  httr::content(type="application/xml") %>% 
-  xml2::as_list() %>%  
-  toJSON(pretty = TRUE, auto_unbox=TRUE) #%>%
-#  fromJSON()
-
-## Read tidy version instead
+```{r}
+package.json <- read_json(
+"https://raw.githubusercontent.com/heroku/node-js-sample/master/package.json")
+package.json
[…]
