### 2017-07-12 12eb5b0
update tests, line wrapping
