+++ tests/testthat/test-crosswalk.R
@@ -3,7 +3,8 @@ testthat::context("crosswalk.R")
 testthat::test_that("we can call crosswalk", {
 
 
-  f <- read_json(system.file("examples/github_format.json", package = "codemetar"))
+  f <- read_json(system.file("examples/github_format.json",
+                             package = "codemetar"))
   a <- crosswalk(f, "GitHub")
   testthat::expect_is(a, "json")
   writeLines(a, "test.json")
@@ -19,7 +20,15 @@ testthat::test_that("we can call crosswalk", {
   testthat::expect_true(v)
   unlink("nodejs.json")
 
+  ## Test add and drop context
   b <- drop_context(a)
   add_context(b, "https://doi.org/doi:10.5063/schema/codemeta-2.0")
+
+
+  ## Test transforms between columns
+  f <- read_json(system.file("examples/github_format.json",
+                             package = "codemetar"))
+  crosswalk(f, "GitHub", "Zenodo")
+
 
 })
