### 2017-07-12 b49e665
Crosswalk (#21)
* Update crosswalk function:

- Function now performs crosswalk on json input.
- Function modified to accomidate explicit type declarations found in codemeta context (and schema.org).
- Unit testing of crosswalk function updated to check validity of output as codemeta

- update vignette to reflect changes (still could use more text explaining use cases, should include more examples, e.g. with package.json, and should illustrate validation check).

* update docs, avoid gh suggests

* update pkgdown site

* update tests, line wrapping
