+++ R/crosswalk.R
@@ -1,62 +1,143 @@
-
-
-
-#' Crosswalk a column to return a JSON-LD context for it based on codemeta
+#' crosswalk
 #'
-#' @param column column name from the CrossWalk table (character vector)
-#' @param cw_table A CrossWalk table. call to
-#' readr::read_csv("https://github.com/codemeta/codemeta/raw/master/crosswalk.csv")
-#'  or use default cached copy.
-#' @param to_json logical, should the result be turned into a json string?
-#'  Default FALSE, which permits easier manipulation such as adding this
-#'  context to an existing list before serializing to JSON
-#' @return a context list, which can be converted to a json string
+#' @param x a JSON list of data fields to be crosswalked
+#' @param from the corresponding column name from the crosswalk table.
+#' @param to the column to translate into, assumes "codemeta" by default
+#' @param codemeta_context the address or contents of codemeta context. Leave at default
+#'
+#' @return a `json` object containing a valid codemeta.json file created by crosswalking the input JSON
+#'
 #' @export
-#' @importFrom stats na.omit
 #'
 #' @examples
-#' crosswalk("Zenodo")
-crosswalk <- function(column,
-                      cw_table = NULL,
-                      to_json = FALSE) {
-  if (is.null(cw_table))
-    cw_table <- cw
+#' \dontrun{
+#' ## Crosswalk data returned by the GitHub API into CodeMeta format
+#' r <- gh::gh("/repos/:owner/:repo", owner = "ropensci", repo = "EML")
+#' crosswalk(r, "GitHub")
+#' }
+#'
+crosswalk <- function(x,
+                      from,
+                      to = "codemeta",
+                      codemeta_context =
+                        "https://doi.org/10.5063/schema/codemeta-2.0"
+){
 
-  def <- function(r) {
-    prefix <- strsplit(r[["Parent Type"]], ":")[[1]][[1]]
-    if (prefix == "schema")
-      out <- paste(prefix, r[["Property"]], sep = ":")
-    ## Best to declare type on any property we want to explicitly
-    ## type in the output version (e.g. codemeta objects)
-    ## Otherwise will not de-reference the `codemeta:` prefix
-    else if (prefix == "codemeta") {
-      type <- gsub("(\\w+).*", "\\1", r[["Type"]])
-      out <- list(
-        "@id" = paste0(prefix, ":", r[["Property"]]),
-        "@type" = paste0("http://schema.org/", type)
-      )
-    }
-    out
-  }
-
-  df <-
-    stats::na.omit(cw_table[c("Parent Type", "Property", "Type", column)])
-
-  ## apply by row
-  context <- apply(df, 1, def)
-  names(context) <- gsub("(\\w+).*", "\\1", df[[column]])
-
-  context <- c(
-    list(schema = "http://schema.org/",
-         codemeta = "https://codemeta.github.io/terms/"),
-    as.list(context)
-  )
-  context
-  if (to_json) {
-    jsonlite::toJSON(list(context = context),
-                     pretty = TRUE,
-                     auto_unbox = TRUE)
+  from_context <- crosswalk_context(crosswalk_table(from), codemeta_context)
+  if(to != "codemeta"){
+    to_context <- crosswalk_context(crosswalk_table(to),  codemeta_context)
+    to_context <- toJSON(to_context, auto_unbox = TRUE, pretty = TRUE)
   } else {
-    context
+    to_context <- codemeta_context
   }
+  crosswalk_transform(x,
+                      crosswalk_context = from_context,
+                      codemeta_context = to_context)
+
 }
+
+# return a subset of the crosswalk table containing codemeta properties and matching column
+#' @importFrom readr read_csv cols
[…]
