+++ vignettes/translating.Rmd
@@ -24,6 +24,8 @@ library("httr")
 library("readr")
 ```
 
+## JSON-LD transforms: Expansion and Compaction
+
 One of the central motivations of JSON-LD is making it easy to translate between different representations of what are fundamentally the same data types. Doing so uses the two core algorithms of JSON-LD: *expansion* and *compaction*, as [this excellent short video by JSON-LD creator Manu Sporny](https://www.youtube.com/watch?v=Tm3fD89dqRE) describes.
 
 Here's how we would use JSON-LD (from R) to translate between the two examples of JSON data from different providers as shown in the video.  First, the JSON from the original provider:
@@ -59,200 +61,80 @@ jsonld_expand(ex) %>%
   jsonld_compact(context = bighash_context)
 ```
 
-## Crosswalk contexts
+## Crosswalking
 
-The CodeMeta Crosswalk table seeks to accomplish a very similar goal.  The crosswalk table provides a human-readable mapping of different software metadata providers into the codemeta context (an extension of schema.org).  
+The CodeMeta Crosswalk table seeks to accomplish a very similar goal.  The crosswalk table provides a human-readable mapping of different software metadata providers into the codemeta context (an extension of schema.org).  For instance, we'll read in some data from GitHub:
 
-We'll start with the actual crosswalk table itself:
 
-```{r message=FALSE}
-crosswalk <- "https://github.com/codemeta/codemeta/raw/master/crosswalk.csv"
-cw <- readr::read_csv(crosswalk)
+### GitHub
+
+Here we crosswalk the JSON data returned as the repository information from the GitHub API: 
+
+```{r eval = FALSE}
+repo_info <- gh::gh("/repos/:owner/:repo", owner = "ropensci", repo = "EML")
 ```
 
-
-## DataCite
-
-
-We can illustrate using the expansion and compaction process to crosswalk DataCite's metadata schema into other formats, such as the native codemeta format or the Zenodo format.  First, let's start with some DataCite metadata:
-
-
-```{r include=FALSE, eval=FALSE}
-## nope, sadly this is not quite clean json translation 
-
-httr::GET("https://doi.org/10.5281/zenodo.573741", 
-    httr::add_headers(Accept="application/vnd.datacite.datacite+xml")) %>% 
-  httr::content(type="application/xml") %>% 
-  xml2::as_list() %>%  
-  toJSON(pretty = TRUE, auto_unbox=TRUE)
-
-## Read tidy version instead
+```{r include = FALSE}
+## Actually, we'll use a chached copy and not eval the above chunk to avoid a `gh` dependency:
+repo_info <- read_json(system.file("examples/github_format.json", package = "codemetar"))
 ```
 
-
+Let's just take a look at what the returned json data looks like:
 
 ```{r}
-datacite_ex <- "https://raw.githubusercontent.com/codemeta/codemetar/master/inst/examples/datacite-xml.json"
-cat(readLines(datacite_ex), sep="\n")
+repo_info
 ```
 
-Note this example uses a JSON-ified representation of DataCite's internal schema, which is XML-based.  So far, this is just plain JSON, with no context.  We will use the crosswalk table to define the DataCite context with reference to codemeta:
-
+We can crosswalk this information into codemeta just by supplying the column name to the `crosswalk` function.  This performs the same expansion of the metadata in the GitHub context, followed by compaction into the codemeta context.  Note that terms not recognized/included in the codemeta context will be dropped:
 
 ```{r}
-## read the file
-datacite_list <- read_json(datacite_ex)
-
-## add context using the codemeta crosswalk function on the current crosswalk table
[…]
