### 2017-07-14 #130 @maelle
@annakrystalli @toph-allen Thanks a lot for accepting to review this package! Your reviews are due on the 2017-08-07.

As a reminder here is a link to the [reviewing guide](https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md) and here is one to the [review template](https://github.com/ropensci/onboarding/blob/master/reviewer_template.md).

@cboettig you can be one of the first to use test our new dynamic RO package status badge! You can add the following to your README.
