### 2017-07-14 08be9cf
add shiny, dynamic onboarding-progress badge
adds to pkgdown and readme.  thanks @maelle et al!
