+++ README.Rmd
@@ -7,6 +7,8 @@ output: github_document
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/codemeta/codemetar?branch=master&svg=true)](https://ci.appveyor.com/project/cboettig/codemetar)
 [![Coverage Status](https://img.shields.io/codecov/c/github/codemeta/codemetar/master.svg)](https://codecov.io/github/codemeta/codemetar?branch=master)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/codemetar)](https://cran.r-project.org/package=codemetar)
+[![](https://ropensci.org/badges/130_status.svg)](https://github.com/ropensci/onboarding/issues/130)
+
 
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
