### 2017-07-20 codemeta/codemetar#23 @cboettig
Generate review metadata
@noamross Really like your suggestion about including a bit more about the review.  Trying to think how best to do this using existing vocabularies, since it makes interoperability of data so much easier.

```
"review":  {
   "@type": "Review",
   "url": "https://github.com/ropensci/onboarding/issues/130",
   "provider": "http://ropensci.org"
                  }
```

I'm not sure what the right field is to indicate the status of the review (e.g. the `in review`, `accepted` etc), maybe `status` would be the term?  Note that http://schema.org/Review defines the property `reviewRating` but that's obviously not the context we really have in mind here.

I've indicated `ropensci` as the provider (could be more verbose and indicate `ropensci` is an organization), note that it would be natural to include the review author & editor here too, (not clear if you'd list all, and of course that would be harder to scrape from the `svg` badge...)

Wanted to cc @mfenner on this too, Martin, any thoughts at a common vocabulary for describing reviews of scholarly works?  Has this come up at all on your end?
