### 2017-07-25 3a1deca
New rOpenSci badge link (#24)
* Update README.Rmd

* Update README.md
