### 2017-08-13 #130 @toph-allen
> However, documentation is present for a few non-exported functions, and a top-level documentation file is missing.
> Examples are missing from a few functions.

>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

> A local run of `devtools::check()` fails due to the lack of `library(tibble)` in the `codemeta-parsing.Rmd` vignette.

Estimated hours spent reviewing: 3h

---

This package aims to make it easier to add linked metadata to R software projects, which is a really cool endeavor — lowering the barrier of entry to linked metadata means its broader adoption within scientific computing. The package does so mainly by parsing existing metadata specified in the R `DESCRIPTION` file and writing a JSON-LD metadata file containing that and other info. The README.md succinctly points people towards existing documentation of the larger CodeMeta Project and JSON-LD. As someone familiar with JSON-LD and linked data paradigms, but not the CodeMeta project, it took me little time to figure out the package’s functionality. The `crosswalk()` function took a little longer to comprehend. With a few changes, I’d definitely recommend approving this package.

#### Specific Comments

I encountered an error when I first ran `install()` on the cloned package. The dependency package `V8` failed to install. However, the error message was helpful in specifying the version I needed to install with `homebrew`.

Running `build_vignettes()` at first failed, with `Error: processing vignette 'codemeta-parsing.Rmd' failed with diagnostics: could not find function "as_tibble"`. Adding `library(tibble)` to that vignette’s library statements should solve that problem, as building the vignettes after loading that package succeeded. This also caused `devtools::check()` to fail.

Going through the package’s documentation alongside the [ROpenSci Packaging Guide’s Documentation section](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#docs), I made the following observations:

- There’s no top-level documentation. The package should use a `codemetar.R` file with package-level help, so that `?codemetar` leads to something (e.g. https://github.com/tidyverse/dplyr/blob/master/R/dplyr.r).
- Help is present for a function `add_context()` which isn’t exported, whereas sister function `drop_context()` isn’t exported. If one is exported, the other probably should be, and vice versa. If both are exported, they should have examples added to their documentation.
- Help is present for functions `crosswalk()` and `crosswalk_transform()`, the latter of which isn’t exported. The former is an exported function, but isn’t well-documented — the “Description” field is just “crosswalk”, and doesn’t sufficiently explain what the function does. Maybe it’s not supposed to be exported, but if it is, its purpose should be explained more fully.
- The `write_codemeta()` function should have a fuller ”Description” statement, but is otherwise sufficient. Specifically, it should explain how its usage differs from `create_codemeta()`.

The vignettes and README.md file are generally rich, and a great source for expanding a user’s understanding of the package’s purpose and potential use cases. However, for entirely novice users, a reference to some introductory material might be helpful (per the package’s previous review).

	There is no `CONTRIBUTING` file, and the `README.md` doesn’t contain community guidelines. However, the `DESCRIPTION` file specifies the requisite fields.

The package’s core functionality successfully executed when I tested it on a few different packages (i.e. `write_codemeta()` wrote the appropriate JSON-LD file.

The paper.md “References” section is currently empty.

The file `R/add_metadata.R` seems to be a vestige of development, and contains a bit of commented-out code. If it’s not needed, it should probably be deleted.

I’m not sure what the purpose of `R/sysdata.rda` file is (and from a few brief searches of the code, could not find a reference to it). If it serves none, it should probably be removed.

The exported functions `write_codemeta()` and `create_codemeta()` both follow a nice verb-subject naming scheme, and also reference the package name. It seems like the function `codemeta_validate()` could be renamed to `validate_codemeta()` to be consistent with these, though that is just personal preference.

The `crosswalk()` function looks like it requires internet access. This should probably be noted somewhere.
