### 2017-08-19 bbf2024
update exported function descriptions, closes #27
