+++ R/codemeta_validate.R
@@ -1,6 +1,8 @@
 #' codemeta validate
 #'
-#' Perform basic check of JSON-LD expanding and compacting, verifying that the result
+#' Checks that a round-trip of expanding and compacting json-ld matches the
+#' original document.  Incorrect schema terms or types (or spelling errors of properties)
+#' will cause this round-trip to fail by not fully compacting.
 #'
 #' @param codemeta path/filename to a codmeta.json file
 #' @param context URL (or path or json string) for the codemeta context.
