+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(add_context)
 export(codemeta_validate)
 export(create_codemeta)
 export(crosswalk)
