### 2017-08-19 #130 @cboettig
@annakrystalli @toph-allen A huge thanks for these immensely helpful reviews!  It is such a treat to get this kind of detailed attention, and I particularly appreciate the efforts of you both to wrestle not only with implementation details but also with the larger concepts and how they are communicated.

I love @annakrystalli's comment about being a gateway drug to json-ld, it certainly has been for me and several of the vignettes really began as (perhaps still are) my playground for learning json-ld and what we can do with it.  I think software metadata is generally both familiar and simple/limited enough (but not trivially so) to make it a good intro into the larger ideas.

Still working through the individual issues back in the codemetar repo and hope to have them done soon, but just wanted to chime in here while it is not too stale and say thanks!
