### 2017-08-19 d3c4f4d
add better background to intro drop sysdata.rda improve package description in DESCRIPTION
