### 2017-08-19 cc67661
codemeta package level documentation closes #25
