### 2017-08-19 c240adc
update vignette index entry
