### 2017-09-14 31cd263
Create CONTRIBUTING.md
