### 2017-09-14 codemeta/codemetar#30 @cboettig
Toph review
> I encountered an error when I first ran install() on the cloned package. The dependency package V8 failed to install. However, the error message was helpful in specifying the version I needed to install with homebrew.

yup, credit to Jeroen an the V8 package there.

> Running build_vignettes() at first failed, with Error: processing vignette 'codemeta-parsing.Rmd' failed with diagnostics: could not find function "as_tibble". Adding library(tibble) to that vignette’s library statements should solve that problem, as building the vignettes after loading that package succeeded. This also caused devtools::check() to fail.

 Good catch, fixed.

> Going through the package’s documentation alongside the ROpenSci Packaging Guide’s Documentation section, I made the following observations:

> There’s no top-level documentation. The package should use a codemetar.R file with package-level help, so that ?codemetar leads to something (e.g. https://github.com/tidyverse/dplyr/blob/master/R/dplyr.r).

Added now, thanks!  See #25

> Help is present for a function add_context() which isn’t exported, whereas sister function drop_context() isn’t exported. If one is exported, the other probably should be, and vice versa. If both are exported, they should have examples added to their documentation.

Both are now exported and have examples

> Help is present for functions crosswalk() and crosswalk_transform(), the latter of which isn’t exported. The former is an exported function, but isn’t well-documented — the “Description” field is just “crosswalk”, and doesn’t sufficiently explain what the function does. Maybe it’s not supposed to be exported, but if it is, its purpose should be explained more fully.

> The write_codemeta() function should have a fuller ”Description” statement, but is otherwise sufficient. Specifically, it should explain how its usage differs from create_codemeta().

Added to the description. (#26)

> The vignettes and README.md file are generally rich, and a great source for expanding a user’s understanding of the package’s purpose and potential use cases. However, for entirely novice users, a reference to some introductory material might be helpful (per the package’s previous review).

Indeed!  Substantial background material has been added along the lines advised in Anna's review.

> There is no `CONTRIBUTING` file, and the `README.md` doesn’t contain community guidelines.

Added.

> However, the `DESCRIPTION` file specifies the requisite fields.
> The package’s core functionality successfully executed when I tested it on a few different packages (i.e. write_codemeta() wrote the appropriate JSON-LD file.

:+1:

> The paper.md “References” section is currently empty.

The only refs are URLs already in the description, so I've just dropped that section; though would appreciate more clarity on that from anyone more familiar with JOSS.

> The file R/add_metadata.R seems to be a vestige of development, and contains a bit of commented-out code. If it’s not needed, it should probably be deleted.

It's more of a reminder to me for a future feature to create codemeta json-ld file from scratch, rather than extracting it from R package locations. Would potentially be useful to document other types of software.

> I’m not sure what the purpose of R/sysdata.rda file is (and from a few brief searches of the code, could not find a reference to it). If it serves none, it should probably be removed.

Removed (was just a cached copy of the crosswalk, but decided the online update was best).  (#29)

> The exported functions write_codemeta() and create_codemeta() both follow a nice verb-subject naming scheme, and also reference the package name. It seems like the function codemeta_validate() could be renamed to validate_codemeta() to be consistent with these, though that is just personal preference.

Yeah, I was trying to following Hadley on this (e.g. see `xml2`).  The standard advice is that functions go from general to specific (i.e. start with package namespace), which is best for tab completion.  The `read` / `write` functions are usually done backwards just to match R's other read/write functions.  So I guess `create_codemeta` is backwards.

> The `crosswalk()` function looks like it requires internet access. This should probably be noted somewhere.

added to the function's documentation.
