+++ man/write_codemeta.Rd
@@ -26,7 +26,11 @@ to package root. Default guess is current dir.}
 writes out the codemeta.json file
 }
 \description{
-write out a codemeta.json file for a given package.
+write out a codemeta.json file for a given package.  This function
+is basically a wrapper around create_codemeta() to both create the
+codemeta object and write it out to a JSON-LD-formatted file in one command.
+It can also be used simply to write out to JSON-LD any existing object
+created with create_codemeta().
 }
 \details{
 If pkg is a codemeta object, the function will attempt to
