### 2017-09-18 d3b96e8
introduce CRAN-friendly extension mechanism X-schema.org-*
