+++ man/crosswalk_transform.Rd
@@ -8,7 +8,7 @@ crosswalk_transform(x, crosswalk_context = NULL,
   codemeta_context = "https://doi.org/10.5063/schema/codemeta-2.0")
 }
 \arguments{
-\item{x}{a JSON list of data fields to be crosswalked}
+\item{x}{a JSON list or file with data fields to be crosswalked}
 
 \item{crosswalk_context}{Context to be added to x}
 

[…]
