### 2017-09-18 a37cec1
update docs, revisions to crosswalk method
