+++ R/crosswalk.R
@@ -3,7 +3,7 @@
 #' Crosswalk between different metadata fields used by different repositories,
 #' registries and archives. For more details see https://codemeta.github.io/crosswalk
 #' This function requires an internet connection to obtain the latest crosswalk table.
-#' This function essentially applies the crosswalk table shown by \link{\code{crosswalk_table}}
+#' This function essentially applies the crosswalk table shown by \code{\link{crosswalk_table}}
 #' to a given JSON metadata record.
 #' @param x a JSON list or file with data fields to be crosswalked
 #' @param from the corresponding column name from the crosswalk table.
