+++ man/crosswalk.Rd
@@ -23,7 +23,7 @@ a `json` object containing a valid codemeta.json file created by crosswalking th
 Crosswalk between different metadata fields used by different repositories,
 registries and archives. For more details see https://codemeta.github.io/crosswalk
 This function requires an internet connection to obtain the latest crosswalk table.
-This function essentially applies the crosswalk table shown by \link{\code{crosswalk_table}}
+This function essentially applies the crosswalk table shown by \code{\link{crosswalk_table}}
 to a given JSON metadata record.
 }
 \examples{
