+++ README.md
@@ -36,6 +36,8 @@ write_codemeta("testthat")
 write_codemeta(".")
 ```
 
+Which creates a file looking like so (first 10 lines; see full [codemeta.json here](https://github.com/codemeta/codemetar/blob/master/codemeta.json)):
+
     {
       "@context": [
         "https://doi.org/doi:10.5063/schema/codemeta-2.0",
@@ -46,265 +48,6 @@ write_codemeta(".")
       "description": "The 'Codemeta' Project defines a 'JSON-LD' format for describing\n  software metadata, as detailed at <https://codemeta.github.io>. This package\n  provides utilities to generate, parse, and modify 'codemeta.jsonld' files \n  automatically for R packages, as well as tools and examples for working with\n  'codemeta' 'JSON-LD' more generally.",
       "name": "codemetar: Generate CodeMeta Metadata for R Packages",
       "issueTracker": "https://github.com/codemeta/codemetar/issues",
-      "license": "https://spdx.org/licenses/MIT",
-      "version": "0.1.0",
-      "programmingLanguage": {
-        "@type": "ComputerLanguage",
-        "name": "R",
-        "version": "3.4.1",
-        "url": "https://r-project.org"
-      },
-      "runtimePlatform": "R version 3.4.1 (2017-06-30)",
-      "provider": {
-        "@id": "https://cran.r-project.org",
-        "@type": "Organization",
-        "name": "Central R Archive Network (CRAN)",
-        "url": "https://cran.r-project.org"
-      },
-      "author": [
-        {
-          "@type": "Person",
-          "givenName": "Carl",
-          "familyName": "Boettiger",
-          "email": "cboettig@gmail.com",
-          "@id": "http://orcid.org/0000-0002-1642-628X"
-        }
-      ],
-      "copyrightHolder": [
-        {
-          "@type": "Person",
-          "givenName": "Carl",
-          "familyName": "Boettiger",
-          "email": "cboettig@gmail.com",
-          "@id": "http://orcid.org/0000-0002-1642-628X"
-        }
-      ],
-      "maintainer": {
-        "@type": "Person",
-        "givenName": "Carl",
-        "familyName": "Boettiger",
-        "email": "cboettig@gmail.com",
-        "@id": "http://orcid.org/0000-0002-1642-628X"
-      },
-      "softwareSuggestions": [
-        {
-          "@type": "SoftwareApplication",
-          "identifier": "testthat",
-          "name": "testthat",
-          "provider": {
-            "@id": "https://cran.r-project.org",
-            "@type": "Organization",
-            "name": "Central R Archive Network (CRAN)",
-            "url": "https://cran.r-project.org"
-          }
-        },
-        {
-          "@type": "SoftwareApplication",
-          "identifier": "jsonvalidate",
-          "name": "jsonvalidate",
-          "provider": {
-            "@id": "https://cran.r-project.org",
-            "@type": "Organization",
-            "name": "Central R Archive Network (CRAN)",
-            "url": "https://cran.r-project.org"
-          }
-        },
-        {
-          "@type": "SoftwareApplication",
-          "identifier": "covr",
-          "name": "covr",
-          "provider": {
-            "@id": "https://cran.r-project.org",
-            "@type": "Organization",
-            "name": "Central R Archive Network (CRAN)",
-            "url": "https://cran.r-project.org"
-          }
-        },
-        {
-          "@type": "SoftwareApplication",
-          "identifier": "knitr",
-          "name": "knitr",
-          "provider": {
-            "@id": "https://cran.r-project.org",
-            "@type": "Organization",
-            "name": "Central R Archive Network (CRAN)",
-            "url": "https://cran.r-project.org"
-          }
-        },
[…]
