### 2017-09-18 #130 @cboettig
Ok, with apologies for the delay, all edits have now been addressed.  I've tried to summarize changes in issues under an (onboarding milestone](https://github.com/codemeta/codemetar/issues?q=is%3Aissue+is%3Aclosed), there are summaries issue for all of the changes made in response to [Toph](https://github.com/codemeta/codemetar/issues/30) and [Anna](https://github.com/codemeta/codemetar/issues/31).  Thanks again for all the help!
