### 2017-09-19 915978e
CRAN metadata extensions (#32)
* introduce CRAN-friendly extension mechanism X-schema.org-*

* update documentation

* update docs, revisions to crosswalk method
