### 2017-09-19 f04ff5e
add cran comments file
