+++ cran-comments.md
@@ -1,0 +1,26 @@
+Dear CRAN Maintainers,
+
+I am pleased to submit the codemetar package for generating codemeta.json files for R packages. Note that
+this package uses the X-schema.org-* pattern as previously discussed abd recommended by CRAN maintainers (Kurt Hornik, Aug 27),
+please let me know if this needs to be revised in any way. Thanks!
+
+Sincerely,
+Carl
+
+
+## Test environments
+* local OS X install, R 3.4.1
+* ubuntu 12.04 (on travis-ci), R 3.4.1
+* win-builder (devel and release)
+
+## R CMD check results
+
+0 errors | 0 warnings | 0 notes
+
+* This is a new release.
+
+## Reverse dependencies
+
+This is a new release, so there are no reverse dependencies.
+
+
