### 2017-09-19 #130 @noamross
Thanks, @cboettig! @toph-allen and @annakrystalli, please look over the changes and let us know if they address your concerns. If yes, just say so and check the appropriate boxes up on your review.  If not, let us know that else @cboettig should follow up on.
