### 2017-09-30 c10c395
fix bug in X-schema.org notation
