+++ R/codemeta_description.R
@@ -87,7 +87,7 @@ codemeta_description <-
     ## add any additional codemeta terms found in the DESCRIPTION metadata
     for(term in additional_codemeta_terms){
       ## in DESCRIPTION, these terms must be *prefixed*:
-      X_term <- paste("X-schema.org-", term)
+      X_term <- paste0("X-schema.org-", term)
       if(!is.null(descr[[X_term]])){
         codemeta[[term]] <- gsub("\\s+", "",
                                  strsplit(descr[[X_term]], ",")[[1]])
