### 2017-09-30 7cb883c
isPartOf is more semantically correct than affiliation
