### 2017-10-09 e562ba3
Update info about ORCiD/CRAN. Rmd not compiled.
