### 2017-10-10 5359bcb
named comment also works, as suggested here: https://twitter.com/AchimZeileis/status/917319708713668608
