### 2017-10-16 #130 @toph-allen
I'm sorry for the delay on this. I've looked through @cboettig's response, plus relevant updates to code & documentation, and it all looks great. I give a hearty 👍.

@cboettig, I really appreciate the specific & detailed responses. I'm happy to have been a part of the review process. Please let me know if there is anything else needed from me to proceed!
