### 2017-10-16 #130 @noamross
[I'm wrapping this up as @maelle is on vacation]

Thank you for your follow-ups @toph-allen and @annakrystalli!  @cboettig, a few last things I am finding in final checks. I'll accept pending these changes:

-  A warning in your tests:
```
Warnings ------------------------------------------------------------------------------
1. we can generate codemeta
                    from the root directory of R source code on github (@test-cloned-repo.R#40) - 'devtools::use_build_ignore' is deprecated.
Use 'usethis::use_build_ignore()' instead.
```
-  In the first vignette (https://github.com/codemeta/codemetar/blob/master/vignettes/A-codemeta-intro.Rmd), there's some text left over from reviewer comments in the second section

---

Fix these and we can move on to onboarding.  Here's the checklist for after:

To-dos:
```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
```
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. (We'll turn on the services on our end as needed)
- [ ] Submit to JOSS using the Zenodo DOI.  We will tag it for expedited review.
