### 2017-10-16 bce72b0
correct minor typos
