### 2017-10-17 ffa40bb
Migrate to devtools -> usethis, fix typo
closes #37
