+++ README.Rmd
@@ -97,3 +97,10 @@ See the [DESCRIPTION](https://github.com/codemeta/codemetar/blob/master/DESCRIPT
 ## Going further
 
 Check out all the [codemetar vignettes](https://codemeta.github.io/codemetar/articles/index.html) for tutorials on other cool stuff you can do with codemeta and json-ld.  
+
+
+
+
+
+[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
+
