+++ DESCRIPTION
@@ -25,7 +25,8 @@ Imports: jsonlite (>= 1.3),
     methods,
     stats,
     stringi,
-    readr
+    readr,
+    usethis
 Suggests: testthat,
     jsonvalidate,
     covr,
