### 2017-10-17 4776da7
fix argument name in migrating from devtools::use_build_ignore to usethis::use_build_ignore
