### 2017-10-17 #130 @cboettig
@noamross Thanks!  Looks like `usethis` is not yet on CRAN.  I think you must be running GitHub versions of devtools and usethis to get that warning?

Shall I go ahead and do the transfer and those subsequent steps now?
