+++ NAMESPACE
@@ -8,6 +8,7 @@ export(crosswalk_table)
 export(drop_context)
 export(write_codemeta)
 importFrom(devtools,build)
+importFrom(devtools,use_build_ignore)
 importFrom(git2r,branches)
 importFrom(git2r,repository)
 importFrom(jsonld,jsonld_compact)
@@ -20,7 +21,6 @@ importFrom(methods,is)
 importFrom(readr,cols)
 importFrom(readr,read_csv)
 importFrom(stringi,stri_trans_general)
-importFrom(usethis,use_build_ignore)
 importFrom(utils,as.person)
 importFrom(utils,citation)
 importFrom(utils,installed.packages)
