+++ DESCRIPTION
@@ -25,8 +25,7 @@ Imports: jsonlite (>= 1.3),
     methods,
     stats,
     stringi,
-    readr,
-    usethis
+    readr
 Suggests: testthat,
     jsonvalidate,
     covr,
