### 2017-10-17 #130 @noamross
Ah, I am using development devtools.  Disregard.  Yes, transfer, but please do fix the text in the vignette first.
