### 2017-10-17 57caa46
fix extraneous lines in vignette A
