+++ vignettes/A-codemeta-intro.Rmd
@@ -47,10 +47,6 @@ For more general information about the CodeMeta Project for defining software me
 
 - context: No one likes typing out long URLs all the time.  So instead, the *context* of a JSON-LD file (`"@context"` element) gives us the context for the terms we use, that is, the root URL.  Usually schema.org but domain specific ones also (eg codemeta)
 
-Briefly explain the difference between the data types (ie json, codemeta json-ld, codemeta r list) so that users can be cognisant of them when using package.
-
-Describe how project is the convergence of a number of initiatives:
-
 - [Schema.org](http://schema.org): A major initiative led by Google and other search engines to define a simple and widely used context to link data on the web through a catalogue of standard metadata fields
 
 - [The CodeMeta Project](http://codemeta.github.io): an academic led community initiative to formalise the metadata fields included in typical software metadata records and introduce important fields that did not have clear equivalents. The codemeta [crosswalk](https://codemeta.github.io/crosswalk) provides an explicit map between the metadata fields used by a broad range of software repositories, registries and archives
