### 2017-10-17 #130 @cboettig
@noamross thanks!  Okay, I htink I've fixed that first vignette now as well (https://github.com/codemeta/codemetar/commit/57caa466f25724be5b2b9bc04e53829831304d49).  Transferring now.
