### 2017-10-17 c92623f
migrate badge urls -> ropensci
