### 2017-10-17 #130 @noamross
Approved!  Thanks @toph-allen and @annakrystalli for your reviews and @cboettig for your work.
