### 2017-10-17 #130 @sckott
travis and codecov seem to be good, appveyor is tied to users, not github accounts, so you can add to yours Carl, I also just added to mine if you want to work off that one https://ci.appveyor.com/project/sckott/codemetar
