### 2017-10-17 3546058
update codemeta.json to ropensci urls.  use personal appveyor
