### 2017-11-01 14d2bdc
more robust handling of remote doi context?
