### 2017-11-04 #130 @cboettig
Seems the badge at https://github.com/ropensci/codemetar still says "Under Review", but the badge at https://ropensci.github.io/codemetar/ says "Peer Reviewed" (at least for me).   Looks like  a caching issue -- can we change the expiration time on the badge for the browser cache?
