### 2017-11-04 #130 @noamross
Pinging @karthik. I think we finally fixed the issue of _github_ caching the badges, but I guess this is another header variable to change?
