+++ R/codemeta_description.R
@@ -8,7 +8,8 @@
 
 ## Supporting old versions will be a nuciance
 new_codemeta <- function() {
-  list(`@context` = "https://doi.org/doi:10.5063/schema/codemeta-2.0",
+  list(`@context` = "http://purl.org/codemeta/2.0",
+         #"https://doi.org/doi:10.5063/schema/codemeta-2.0",
   `@type` = "SoftwareSourceCode")
 }
 
