+++ vignettes/C-validation-in-json-ld.Rmd
@@ -42,7 +42,7 @@ Consider the following codemeta document:
 codemeta <- 
 '
 {
-  "@context": "https://doi.org/doi:10.5063/schema/codemeta-2.0",
+  "@context": "http://purl.org/codemeta/2.0",
   "@type": "SoftwareSourceCode",
   "name": "codemetar: Generate CodeMeta Metadata for R Packages",
   "datePublished":" 2017-05-20",
@@ -60,6 +60,11 @@ codemeta <-
 '
 ```
 
+
+```{r}
+jsonld::jsonld_compact(codemeta, "http://purl.org/codemeta/2.0")
+```
+
 ## Framing: subsetting data
 
 By default, frames return all the input data, while our application may only be interested in some subset.  Often it is sufficient just to ignore these additional terms: in the example above it's just as easy for our application to work with author elements whether or not we have dropped other elements such as `meta$version`.  To restrict a frame to returning only the nodes we explicitly mention, we can use the keyword `@explicit`:
@@ -67,8 +72,10 @@ By default, frames return all the input data, while our application may only be 
 
 
 ```{r}
+options(jsonld_use_accept = TRUE)
+
 frame <- '{
-  "@context": "https://doi.org/doi:10.5063/schema/codemeta-2.0",
+  "@context": "http://purl.org/codemeta/2.0",
   "@explicit": "true",
   "@type": "Person",
   "givenName": {},
@@ -106,9 +113,10 @@ We can use a frame with the special directive `"@embed": "@always"` to say that 
 
 ```{r}
 frame <- '{
-  "@context": "https://doi.org/doi:10.5063/schema/codemeta-2.0",
+  "@context": "http://purl.org/codemeta/2.0",
   "@embed": "@always"
 }'
+
 
 meta <- 
   jsonld_frame(codemeta, frame) %>%
@@ -138,7 +146,7 @@ that declares that the `buildInstructions` are included as text, which differs f
 codemeta <- 
 '
 {
-  "@context": "https://doi.org/doi:10.5063/schema/codemeta-2.0",
+  "@context": "https://purl.org/codemeta/2.0",
   "@type": "SoftwareSourceCode",
   "name": "codemetar: Generate CodeMeta Metadata for R Packages",
   "buildInstructions": { 
@@ -153,12 +161,12 @@ When we perform a framing or compaction operation, `buildInstructions` gets de-r
 
 ```{r}
 meta <-
-  jsonld_frame(codemeta, '{"@context": "https://doi.org/doi:10.5063/schema/codemeta-2.0"}') %>% 
+  jsonld_frame(codemeta, '{"@context": "https://purl.org/codemeta/2.0"}') %>% 
   fromJSON(codemeta, simplifyVector = FALSE) %>%  
   getElement("@graph") %>% getElement(1)    
 
 ## above is same as compacting:
-#jsonld_compact(codemeta, "https://doi.org/doi:10.5063/schema/codemeta-2.0") %>% 
+#jsonld_compact(codemeta, "http://purl.org/codemeta/2.0") %>% 
 #  fromJSON(codemeta, simplifyVector = FALSE)
 
 meta$buildInstructions
