+++ DESCRIPTION
@@ -12,8 +12,8 @@ Description: The 'Codemeta' Project defines a 'JSON-LD' format for describing
   automatically for R packages, as well as tools and examples for working with
   'codemeta' 'JSON-LD' more generally.
 License: MIT + file LICENSE
-URL: https://github.com/codemeta/codemetar
-BugReports: https://github.com/codemeta/codemetar/issues
+URL: https://github.com/ropensci/codemetar
+BugReports: https://github.com/ropensci/codemetar/issues
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
