### 2017-11-08 f63a193
update docs, use options(codemeta_context = ) to set context, update pkgdown site
