### 2017-11-10 eede656
also construct orcid url from orcid id
