### 2017-11-10 43b35b4
test should use getOption for context too
