### 2017-11-11 #130 @maelle
@cboettig you haven't submitted the package to JOSS yet, have you?
