### 2017-11-11 #130 @cboettig
@maelle no, thanks for the ping, I'll try that now.  anything special I need to do re connecting to the rOpenSci review?
