### 2017-11-11 #130 @cboettig
okay, I believe I've initiated submission with http://joss.theoj.org/papers/b16b968425f513da31632b0edd8ff0c5.
