### 2017-11-11 #130 @maelle
Perfect, I'll watch the new submissions at JOSS to comment in the Github issue right away to be sure it's flagged as rOpenSci-reviewed.
