### 2017-11-13 5df3df3
Update paper.md (#39)
