+++ paper.md
@@ -1,6 +1,11 @@
 ---
 title: 'Generating CodeMeta Metadata for R Packages'
-tags: ["metadata", "codemeta", "ropensci", "citation", "credit"],
+tags:
+ - metadata
+ - codemeta
+ - ropensci
+ - citation
+ - credit
 authors:
  - name: Carl Boettiger
    orcid: 0000-0002-1642-628X
@@ -14,7 +19,7 @@ bibliography: paper.bib
 
 # Summary
 
-The CodeMeta project, <codemeta.github.io>, defines a 'JSON-LD' format for describing software 
+The CodeMeta project, <codemeta.github.io>, defines a 'JSON-LD' format for describing software
 metadata, based largely on `schema.org` terms.
 This metadata format is being adopted by many leading archives for scientific software, including DataCite,
 Zenodo, and DataONE to address many of the needs identified in the NIH report on the need for a
@@ -25,8 +30,6 @@ The `codemetar` package provides utilities to generate and validate these `codem
 files automatically for R packages by parsing the DESCRIPTION file
 and other common locations for R metadata.
 The package also includes utilities and examples for parsing and working with existing codemeta files,
-and includes several vignettes which illustrate both the basic usage of the package
-as well as some more advanced applications. 
+and includes several vignettes which illustrate both the basic usage of the package as well as some more advanced applications.
 
-
-
+# References
