### 2017-11-13 6fb5145
Treat links as citations as @arfon suggests
