+++ paper.md
@@ -19,13 +19,12 @@ bibliography: paper.bib
 
 # Summary
 
-The CodeMeta project, <codemeta.github.io>, defines a 'JSON-LD' format for describing software
+The CodeMeta project defines a JSON-LD format [@jsonld] for describing software
 metadata, based largely on `schema.org` terms.
 This metadata format is being adopted by many leading archives for scientific software, including DataCite,
 Zenodo, and DataONE to address many of the needs identified in the NIH report on the need for a
-"Software Discovery Index", <http://www.softwarediscoveryindex.org/>.
-Many common software metadata formats have been mapped into CodeMeta by means of a crosswalk table,
-<https://codemeta.github.io/crosswalk/>, also implemented in this package.
+"Software Discovery Index" [@SDI].
+Many common software metadata formats have been mapped into CodeMeta by means of a crosswalk table [@codemetar], also implemented in this package.
 The `codemetar` package provides utilities to generate and validate these `codemeta.json`
 files automatically for R packages by parsing the DESCRIPTION file
 and other common locations for R metadata.
